Move AWS classes used for Glue metastore
