+++ plugin/trino-hive/src/main/java/io/trino/plugin/hive/metastore/glue/AwsApiCallStats.java
@@ -11,7 +11,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package io.trino.plugin.hive.aws;
+package io.trino.plugin.hive.metastore.glue;
 
 import com.google.errorprone.annotations.ThreadSafe;
 import io.airlift.stats.CounterStat;
