+++ plugin/trino-hive/src/main/java/io/trino/plugin/hive/metastore/glue/AwsSdkUtil.java
@@ -14,7 +14,6 @@
 package io.trino.plugin.hive.metastore.glue;
 
 import com.google.common.collect.AbstractIterator;
-import io.trino.plugin.hive.aws.AwsApiCallStats;
 
 import java.util.Iterator;
 import java.util.function.BiConsumer;
