+++ plugin/trino-hive/src/main/java/io/trino/plugin/hive/metastore/glue/GlueMetastoreStats.java
@@ -14,8 +14,6 @@
 package io.trino.plugin.hive.metastore.glue;
 
 import com.amazonaws.metrics.RequestMetricCollector;
-import io.trino.plugin.hive.aws.AwsApiCallStats;
-import io.trino.plugin.hive.aws.AwsSdkClientCoreStats;
 import org.weakref.jmx.Flatten;
 import org.weakref.jmx.Managed;
 import org.weakref.jmx.Nested;
