+++ plugin/trino-iceberg/src/test/java/io/trino/plugin/iceberg/catalog/glue/GlueMetastoreMethod.java
@@ -14,7 +14,7 @@
 package io.trino.plugin.iceberg.catalog.glue;
 
 import com.google.common.math.DoubleMath;
-import io.trino.plugin.hive.aws.AwsApiCallStats;
+import io.trino.plugin.hive.metastore.glue.AwsApiCallStats;
 import io.trino.plugin.hive.metastore.glue.GlueMetastoreStats;
 
 import java.math.RoundingMode;
